Unify OsString/OsStr for byte-based implementations
