+++ src/libstd/sys/redox/mod.rs
@@ -22,7 +22,6 @@ pub mod memchr;
 pub mod mutex;
 pub mod net;
 pub mod os;
-pub mod os_str;
 pub mod path;
 pub mod pipe;
 pub mod process;
@@ -34,6 +33,8 @@ pub mod syscall;
 pub mod thread;
 pub mod thread_local;
 pub mod time;
+
+pub use crate::sys_common::os_str_bytes as os_str;
 
 #[cfg(not(test))]
 pub fn init() {}
